# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Auth & Payment API
  description: |
    會員登入 + 付款方式綁定。BE 與 FE 共用此契約。
    任何端點變動必須先改此檔，再開發。
  version: 0.1.0
  contact:
    name: AI Dev Workflow Demo
servers:
  - url: https://api.dev.example.com
    description: Development
  - url: https://api.staging.example.com
    description: Staging
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Email 註冊
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 註冊成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email 已被註冊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                ok: false
                error: { code: USER_EXISTS, message: "Email already registered" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 輸入驗證失敗
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, minLength: 10, description: "至少 10 字，含英數" }
...
    AuthSuccess:
      type: object
      properties:
        ok: { type: boolean, const: true }
        data:
          type: object
          properties:
            access_token: { type: string }
            expires_in:   { type: integer }
            user:         { $ref: '#/components/schemas/User' }
    User:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        email:               { type: string, format: email }
        two_fa_enabled:      { type: boolean }
        has_payment_method:  { type: boolean }
        created_at:          { type: string, format: date-time }
...
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: WEAK_PASSWORD }
            message: { type: string }